Adicionar produtos adicionado

+++ MauiAppMinhasCompras/Views/NovoProduto.xaml.cs
@@ -16,12 +16,13 @@ public partial class NovoProduto : ContentPage
             Produto p = new Produto
             {
                 Descricao = txt_descricao.Text,
-                Quantidade = Convert.ToInt32(txt_quantidade.Text),
+                Quantidade = Convert.ToDouble(txt_quantidade.Text),
                 Preco = Convert.ToDouble(txt_preco.Text)
             };
 
             await App.DB.Insert(p);
             await DisplayAlert("Sucesso!", "Registro inserido", "OK");
+
         }
         catch (Exception ex)
         {
